회원 관리

+++ FinalProject/src/main/webapp/WEB-INF/views/manage/user/user/allApprobate.jsp
@@ -23,28 +23,27 @@
 					<th>회원 검색</th>
 					<td colspan="">
 						<span class="sbox small">
-							<select name="searchType">
+							<select name="searchType" id="searchType">
 								<option value="1">아이디</option>
 								<option value="2">회원명</option>
-								<option value="3">휴대폰번호</option>
 							</select>
 						</span>
-						<input class="tbox" name="searchWord" value="">
+						<input class="tbox" name="searchWord" id="searchWord" value="">
 					</td>
 				</tr>
 				<tr>
 					<th>회원 등급</th>
 					<td colspan="3">
 						<label>
-							<input type="checkbox" class="checkUserType">
+							<input type="checkbox" name="user" value="일반회원" class="checkUserType">
 							<span>일반회원</span>
 						</label>
 						<label class="ml10">
-							<input type="checkbox" class="checkUserType">
+							<input type="checkbox" name="user" value="운송기사" class="checkUserType">
 							<span>운송기사</span>
 						</label>
 						<label class="ml10">
-							<input type="checkbox" class="checkUserType">
+							<input type="checkbox" name="user" value="부동산" class="checkUserType">
 							<span>부동산</span>
 						</label>
 					</td>
@@ -140,13 +139,17 @@
 </body>
 </html>
 <script>
-	
+
 $(".checkUserType").change(function(){
-	var bool = $(this).is(":checked");
-	var type = $(this).next().html();
+	var bool = $("input[name='user']:checked").length;
+	var val = "";
+	for(var i = 0; i<bool; i++){
+		val += $("input[name='user']:checked").eq(i).val()+",";
+	}
+	var type = val.substring(0,val.length-1);
 	console.log(type);
 	console.log(bool);
-	if(bool){
+	if(bool==1){
 		$.ajax({
 			type: "post",
 			url: "/userManage/userCheckList",
@@ -155,7 +158,98 @@ $(".checkUserType").change(function(){
 				$(".list_table").children().eq(2).html("");
 				var listV = $(".list_table").children().eq(2);
 				console.log(listV);
-				str = "";
+				var str="";
+				for(var i=0; i<data.list.length; i++){
+					str += "<tr>";
+					str += "<td>"+data.list[i].userId+"</td>";
+					str += "<td>"+data.list[i].userPw+"</td>";
+					str += "<td>"+data.list[i].userName+"</td>";
+					str += "<td>"+data.list[i].userAddr+"</td>";
+					str += "<td>"+data.list[i].userPhone+"</td>";
+					str += "<td>"+data.list[i].userEmail+"</td>";
+					str += "<td>"+data.list[i].type+"</td>";
+					str += "<td>"+data.list[i].enrollDate+"</td>";
+					str += "<td><a href='/userManage/deleteUser?userIdx='"+data.list[i].userIdx+" "+"class='tag col_blue f_w'>";
+					str += "회원탈퇴</a></td>";
+					str += "</tr>";
+					
+				}
+				listV.append(str);
+			}
+		});
+	}else if(bool==2){
+		$.ajax({
+			type: "post",
+			url: "/userManage/userCheckList",
+			data: {type:type},
+			success: function(data){
+				$(".list_table").children().eq(2).html("");
+				var listV = $(".list_table").children().eq(2);
+				console.log(listV);
+				var str ="";
+				for(var i=0; i<data.list.length; i++){
+					str += "<tr>";
+					str += "<td>"+data.list[i].userId+"</td>";
+					str += "<td>"+data.list[i].userPw+"</td>";
+					str += "<td>"+data.list[i].userName+"</td>";
+					str += "<td>"+data.list[i].userAddr+"</td>";
+					str += "<td>"+data.list[i].userPhone+"</td>";
+					str += "<td>"+data.list[i].userEmail+"</td>";
+					str += "<td>"+data.list[i].type+"</td>";
+					str += "<td>"+data.list[i].enrollDate+"</td>";
+					str += "<td><a href='/userManage/deleteUser?userIdx='"+data.list[i].userIdx+" "+"class='tag col_blue f_w'>";
+					str += "회원탈퇴</a></td>";
+					str += "</tr>";
+					
+				}
+				listV.append(str);
+			}
+		});
+	}else if(bool==0){ 
+		$.ajax({
+			type: "post",
+			url: "/userManage/approbate",
+			success: function(data){
+				$(".list_table").children().eq(2).html("");
+				var listV = $(".list_table").children().eq(2);
+				console.log(listV);
+				console.log(data.list);
+				var str2 = "";
+				for(var i=0; i<data.list.length; i++){
+					var type= "";
+					if(data.list[i].userType == "1"){
+						type = "운송기사";
+					}else if(data.list[i].userType == "2"){
+						type = "부동산";
+					}else if(data.list[i].userType == "3"){
+						type = "일반회원";
+					}
+					str2 += "<tr>";
+					str2 += "<td>"+data.list[i].userId+"</td>";
+					str2 += "<td>"+data.list[i].userPw+"</td>";
+					str2 += "<td>"+data.list[i].userName+"</td>";
+					str2 += "<td>"+data.list[i].userAddr+"</td>";
+					str2 += "<td>"+data.list[i].userPhone+"</td>";
+					str2 += "<td>"+data.list[i].userEmail+"</td>";
+					str2 += "<td>"+data.list[i].type+"</td>";
+					str2 += "<td>"+data.list[i].enrollDate+"</td>";
+					str2 += "<td><a href='/userManage/deleteUser?userIdx='"+data.list[i].userIdx+" "+"class='tag col_blue f_w'";
+					str2 += ">회원탈퇴</a></td>";
+					str2 += "</tr>";
+				}
+				listV.append(str2);
+			}
+		});
+	}
+	/* if(bool2){
+		$.ajax({
+			type: "post",
+			url: "/userManage/userCheckList",
+			data: {type:type},
+			success: function(data){
+				$(".list_table").children().eq(2).html("");
+				var listV = $(".list_table").children().eq(2);
+				console.log(listV);
 				for(var i=0; i<data.list.length; i++){
 					str += "<tr>";
 					str += "<td>"+data.list[i].userId+"</td>";
@@ -173,19 +267,17 @@ $(".checkUserType").change(function(){
 				}
 				listV.append(str);
 			}
-		}); /* ajax 끝 */
-	}else{ /* if */
+		}); 
+	}else if(bool==0){ 
 		$.ajax({
 			type: "post",
-			url: "/userManage/userListAll",
+			url: "/userManage/approbate",
 			success: function(data){
 				$(".list_table").children().eq(2).html("");
 				var listV = $(".list_table").children().eq(2);
 				console.log(listV);
 				console.log(data.list);
-				str = "";
-				
-				
+				str2 = "";
 				for(var i=0; i<data.list.length; i++){
 					var type= "";
 					if(data.list[i].userType == "1"){
@@ -195,23 +287,63 @@ $(".checkUserType").change(function(){
 					}else if(data.list[i].userType == "3"){
 						type = "일반회원";
 					}
-					str += "<tr>";
-					str += "<td>"+data.list[i].userId+"</td>";
-					str += "<td>"+data.list[i].userPw+"</td>";
-					str += "<td>"+data.list[i].userName+"</td>";
-					str += "<td>"+data.list[i].userAddr+"</td>";
-					str += "<td>"+data.list[i].userPhone+"</td>";
-					str += "<td>"+data.list[i].userEmail+"</td>";
-					str += "<td>"+type+"</td>";
-					str += "<td>"+data.list[i].enrollDate+"</td>";
-					str += "<td><a href='/userManage/deleteUser?userIdx='"+data.list[i].userIdx+" "+"class='tag col_blue f_w'";
-					str += ">회원탈퇴</a></td>";
-					str += "</tr>";
+					str2 += "<tr>";
+					str2 += "<td>"+data.list[i].userId+"</td>";
+					str2 += "<td>"+data.list[i].userPw+"</td>";
+					str2 += "<td>"+data.list[i].userName+"</td>";
+					str2 += "<td>"+data.list[i].userAddr+"</td>";
+					str2 += "<td>"+data.list[i].userPhone+"</td>";
+					str2 += "<td>"+data.list[i].userEmail+"</td>";
+					str2 += "<td>"+type+"</td>";
+					str2 += "<td>"+data.list[i].enrollDate+"</td>";
+					str2 += "<td><a href='/userManage/deleteUser?userIdx='"+data.list[i].userIdx+" "+"class='tag col_blue f_w'";
+					str2 += ">회원탈퇴</a></td>";
+					str2 += "</tr>";
+				}
+				listV.append(str2);
+			}
+		});
+	} *//* else{
+		$.ajax({
+			type: "post",
+			url: "/userManage/userCheckList",
+			data: {type:type},
+			success: function(data){
+				$(".list_table").children().eq(2).html("");
+				var listV = $(".list_table").children().eq(2);
+				console.log(listV);
+				console.log(data.list);
+				str = "";
+				for(var i=0; i<data.list.length; i++){
+					var type= "";
+					if(data.list[i].userType == "1"){
+						type = "운송기사";
+					}else if(data.list[i].userType == "2"){
+						type = "부동산";
+					}else if(data.list[i].userType == "3"){
+						type = "일반회원";
+					}
+					if(data.list[i].userType != "type"){
+						str += "<tr>";
+						str += "<td>"+data.list[i].userId+"</td>";
+						str += "<td>"+data.list[i].userPw+"</td>";
+						str += "<td>"+data.list[i].userName+"</td>";
+						str += "<td>"+data.list[i].userAddr+"</td>";
+						str += "<td>"+data.list[i].userPhone+"</td>";
+						str += "<td>"+data.list[i].userEmail+"</td>";
+						str += "<td>"+type+"</td>";
+						str += "<td>"+data.list[i].enrollDate+"</td>";
+						str += "<td><a href='/userManage/deleteUser?userIdx='"+data.list[i].userIdx+" "+"class='tag col_blue f_w'";
+						str += ">회원탈퇴</a></td>";
+						str += "</tr>";
+					}
 				}
 				listV.append(str);
 			}
-		}); /* ajax 끝 */
-	}
+		});
+	} */
+	
+	
 });
 	
 </script>
